Alterações e separações Sign
