+++ android/app/build.gradle
@@ -66,7 +66,7 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
-    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'com.android.support:multidex:1.0.3' //Foi adicionado para utilizar could_firestore.
 }
 
 apply plugin: 'com.google.gms.google-services'
