+++ lib/main.dart
@@ -1,8 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:scoped_model/scoped_model.dart';
+import 'package:serview/models/user_model.dart';
 import 'package:serview/ui/home_page.dart';
-import 'package:serview/ui/sign_page.dart';
+import 'package:serview/ui/signPage/sign_page.dart';
 import 'package:serview/ui/curriculum_page.dart';
 
-void main() => runApp(MaterialApp(
-      home: SignPage(),
-    ));
+void main() => runApp(
+      ScopedModel<UserModel>(
+          model: UserModel(),
+          child: MaterialApp(
+            home: SignPage(),
+          )),
+    );
