+++ lib/models/user_model.dart
@@ -1,0 +1,12 @@
+import 'package:scoped_model/scoped_model.dart';
+
+class UserModel extends Model{
+  
+  bool isLoading = false;
+
+  void signUp() {}
+
+  void signIn() {}
+
+  void recoverPass() {}
+}
