+++ pubspec.yaml
@@ -30,6 +30,7 @@ dependencies:
   firebase_messaging: ^5.1.8
   scoped_model: ^1.0.1
   path_provider: ^1.4.2
+  url_launcher: ^5.2.5
 
 dev_dependencies:
   flutter_test:
